rename package of trending module
