+++ contrib/piasy/trending/src/debug/AndroidManifest.xml
@@ -23,7 +23,7 @@
 -->
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
         xmlns:tools="http://schemas.android.com/tools"
-        package="com.github.piasy.octostars.trending"
+        package="com.github.piasy.octostars.features.trending"
         >
 
     <application
