+++ contrib/piasy/trending/src/main/AndroidManifest.xml
@@ -23,7 +23,7 @@
   -->
 
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-        package="com.github.piasy.octostars.trending"
+        package="com.github.piasy.octostars.features.trending"
         >
 
     <application
