+++ contrib/piasy/trending/src/main/java/com/github/piasy/octostars/features/trending/TrendingComponent.java
@@ -1,0 +1,45 @@
+/*
+ * The MIT License (MIT)
+ *
+ * Copyright (c) 2017 Piasy
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+
+package com.github.piasy.octostars.features.trending;
+
+import com.github.piasy.octostars.di.AppComponent;
+import com.github.piasy.octostars.misc.MiscModule;
+import com.github.piasy.yamvp.dagger2.ActivityScope;
+import dagger.Component;
+
+/**
+ * Created by Piasy{github.com/Piasy} on 15/9/19.
+ *
+ * Di appComponent for splash.
+ */
+@ActivityScope
+@Component(
+        dependencies = AppComponent.class,
+        modules = {
+                MiscModule.class,
+        })
+interface TrendingComponent {
+    void inject(TrendingActivity activity);
+}
