+++ contrib/piasy/trending/src/main/java/com/github/piasy/octostars/features/trending/TrendingPresenter.java
@@ -1,0 +1,54 @@
+/*
+ * The MIT License (MIT)
+ *
+ * Copyright (c) 2017 Piasy
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+
+package com.github.piasy.octostars.features.trending;
+
+import com.github.piasy.bootstrap.base.utils.RxUtil;
+import com.github.piasy.yamvp.dagger2.ActivityScope;
+import com.github.piasy.yamvp.rx.YaRxPresenter;
+import io.reactivex.android.schedulers.AndroidSchedulers;
+import io.reactivex.schedulers.Schedulers;
+import javax.inject.Inject;
+
+/**
+ * Created by Piasy{github.com/Piasy} on 20/09/2016.
+ */
+
+@ActivityScope
+class TrendingPresenter extends YaRxPresenter<TrendingView> {
+    private final TrendingRepo mTrendingRepo;
+
+    @Inject
+    TrendingPresenter(final TrendingRepo trendingRepo) {
+        super();
+        mTrendingRepo = trendingRepo;
+    }
+
+    void loadTrending(final String lang, final @TrendingRepo.Since String since) {
+        addUtilDestroy(mTrendingRepo.trending(lang, since, false)
+                .subscribeOn(Schedulers.io())
+                .observeOn(AndroidSchedulers.mainThread(), true)
+                .subscribe(getView()::showTrending, RxUtil.OnErrorLogger));
+    }
+}
